// src/interfaces/interfaces.update.ts
export interface WearerUpdateData {
  first_name?: string;
  dob?: string;
  ethnicity_id?: string;
  risk_level_id?: string;
  language_id?: string;
  timezone_id?: string;
  hair_colour_id?: string;
  eye_colour_id?: string;
  surname?: string;
  gender_id?: string;
  phone_type?: string;
  mobile?: string;
  country_id?: string;
  email?: string;
  height?: string;
  weight?: string;
  group_id?: string;
  "buddicombobox-1607-inputEl"?: string;
  device_profile_id?: string;
  device_profile_sb_id?: string;
  wearer_type_id?: string;
  ref?: string;
  criminal_records_no?: string;
  vibrate_tag?: string;
  audible_vibrate?: string;
  rule_priority_id?: string;
  clip_panic_number?: string;
  sb_outbound_number?: string;
  smartid_outbound_number?: string;
  view?: string;
  preset_id?: string;
  notes?: string;
  responsible_officer_id?: string;
  category?: string;
  employer_name?: string;
  employer_position?: string;
  employer_date_started?: string;
  employer_additional?: string;
  employer_address_line_1?: string;
  employer_address_line_2?: string;
  employer_address_city?: string;
  employer_address_county?: string;
  employer_address_postcode?: string;
  employer_telephone?: string;
  identifying_marks?: string;
  risk_assessment?: string;
  medical_conditions?: string;
  "buddicombobox-1797-inputEl"?: string;
  ssid?: string;
  ssid_password?: string;
  interpretor_required?: string;
  exclude_from_proximity?: string;
  exclude_from_loitering?: string;
  exclude_from_auto_day_report?: string;
  filtertype?: string;
  filterstatus?: string;
  startdate?: string;
  enddate?: string;
  filtersearch?: string;
  "datefield-2061-inputEl"?: string;
  "datefield-2062-inputEl"?: string;
  type_filter?: string;
  status_filter?: string;
  priority_filter?: string;
  user_filter?: string;
  date_filter?: string;
  requires_followup?: string;
  start_tagging_time?: string;
  end_tagging_time?: string;
  categories?: string;
  rule_template_ids?: string;
}

export interface UpdateWearerRequest {
  wearer_id: string;
  last_edit_user: string;
  last_edit_timestamp: string;
  force: string;
  wearer: WearerUpdateData;
  request_type: string;
  return_type: string;
  method: string;
  c: string;
  csrf_token: string;
}